refactor: rename package and update test function names for consistency
