--- restaurant-service/src/main/kotlin/com/munchies/restaurant/application/usecases/menu/MenuItemUseCases.kt
@@ -1,124 +1,0 @@
-package com.munchies.restaurant.application.usecases.menu
-
-import com.munchies.restaurant.application.UseCase
-import com.munchies.restaurant.domain.aggregate.CategoryId
-import com.munchies.restaurant.domain.aggregate.MenuId
-import com.munchies.restaurant.domain.aggregate.MenuItemId
-import com.munchies.restaurant.domain.repository.MenuRepository
-import com.munchies.restaurant.domain.valueobject.Money
-import com.munchies.restaurant.domain.valueobject.menu.MenuItemDescription
-import com.munchies.restaurant.domain.valueobject.menu.MenuItemName
-import java.math.BigDecimal
-
-data class MenuItemUseCases(
-  val add: AddMenuItemUseCase,
-  val update: UpdateMenuItemUseCase,
-  val remove: RemoveMenuItemUseCase,
-)
-
-data class AddMenuItemCommand(
-  val menuId: String,
-  val categoryId: String,
-  val name: String,
-  val description: String,
-  val price: BigDecimal,
-)
-
-sealed interface AddMenuItemResult {
-  data class Success(val itemId: String) : AddMenuItemResult
-  data object MenuNotFound : AddMenuItemResult
-  data object CategoryNotFound : AddMenuItemResult
-  data class InvalidItem(val error: String) : AddMenuItemResult
-}
-
-class AddMenuItemUseCase(private val menuRepository: MenuRepository) :
-  UseCase<AddMenuItemCommand, AddMenuItemResult> {
-  override suspend operator fun invoke(command: AddMenuItemCommand): AddMenuItemResult {
-    val menu = menuRepository.findById(MenuId(command.menuId))
-      ?: return AddMenuItemResult.MenuNotFound
-
-    val category = menu.getCategory(CategoryId(command.categoryId))
-      ?: return AddMenuItemResult.CategoryNotFound
-
-    return runCatching {
-      val item = category.addItem(
-        MenuItemName.of(command.name),
-        MenuItemDescription.of(command.description),
-        Money(command.price),
-      )
-      menuRepository.save(menu)
-
-      AddMenuItemResult.Success(item.id.value)
-    }.getOrElse { AddMenuItemResult.InvalidItem(it.message ?: "") }
-  }
-}
-
-data class UpdateMenuItemCommand(
-  val menuId: String,
-  val categoryId: String,
-  val itemId: String,
-  val name: String,
-  val description: String,
-  val price: BigDecimal,
-)
-
-sealed interface UpdateMenuItemResult {
-  data object Success : UpdateMenuItemResult
-  data object MenuNotFound : UpdateMenuItemResult
-  data object CategoryNotFound : UpdateMenuItemResult
-  data class InvalidItem(val error: String) : UpdateMenuItemResult
-}
-
-class UpdateMenuItemUseCase(
-  private val menuRepository: MenuRepository,
-) : UseCase<UpdateMenuItemCommand, UpdateMenuItemResult> {
-  override suspend operator fun invoke(command: UpdateMenuItemCommand): UpdateMenuItemResult {
-    val menu = menuRepository.findById(MenuId(command.menuId))
-      ?: return UpdateMenuItemResult.MenuNotFound
-
-    val category = menu.getCategory(CategoryId(command.categoryId))
-      ?: return UpdateMenuItemResult.CategoryNotFound
-
-    return runCatching {
-      category.updateItem(
-        MenuItemId(command.itemId),
-        MenuItemName.of(command.name),
-        MenuItemDescription.of(command.description),
-        Money(command.price),
-      )
-      menuRepository.save(menu)
-      UpdateMenuItemResult.Success
-    }.getOrElse { UpdateMenuItemResult.InvalidItem(it.message ?: "") }
-  }
-}
-
-data class RemoveMenuItemCommand(
-  val menuId: String,
-  val categoryId: String,
-  val itemId: String,
-)
-
-sealed interface RemoveMenuItemResult {
-  data object Success : RemoveMenuItemResult
-  data object MenuNotFound : RemoveMenuItemResult
-  data object CategoryNotFound : RemoveMenuItemResult
-  data class InvalidItem(val error: String) : RemoveMenuItemResult
-}
-
-class RemoveMenuItemUseCase(
-  private val menuRepository: MenuRepository,
-) : UseCase<RemoveMenuItemCommand, RemoveMenuItemResult> {
-  override suspend operator fun invoke(command: RemoveMenuItemCommand): RemoveMenuItemResult {
-    val menu = menuRepository.findById(MenuId(command.menuId))
-      ?: return RemoveMenuItemResult.MenuNotFound
-
-    val category = menu.getCategory(CategoryId(command.categoryId))
-      ?: return RemoveMenuItemResult.CategoryNotFound
-
-    return runCatching {
-      category.removeItem(MenuItemId(command.itemId))
-      menuRepository.save(menu)
-      RemoveMenuItemResult.Success
-    }.getOrElse { RemoveMenuItemResult.InvalidItem(it.message ?: "") }
-  }
-}
